Add enhanced mappings and services for wishlist and followers
Introduced extended mapping for wish list and follower DTOs, leveraging presigned URLs for avatar and logo retrieval. Updated the controller and repository logic to handle extended wish list data and improve follower handling. Minor cleanups and authorization updates were also implemented.

+++ src/Api/Forja.API/Controllers/UserProfile/UserWishListController.cs
@@ -10,18 +10,25 @@ public class UserWishListController : ControllerBase
 {
     private readonly IUserWishListService _userWishListService;
     private readonly IAuditLogService _auditLogService;
+    private readonly IUserService _userService;
+    private readonly IDistributedCache _cache;
 
     public UserWishListController(IUserWishListService userWishListService,
-        IAuditLogService auditLogService)
+        IAuditLogService auditLogService,
+        IUserService userService,
+        IDistributedCache cache)
     {
         _userWishListService = userWishListService;
         _auditLogService = auditLogService;
+        _userService = userService;
+        _cache = cache;
     }
 
     /// <summary>
     /// Retrieves all UserWishList entries.
     /// </summary>
     /// <returns>A list of UserWishListDTO.</returns>
+    [Authorize(Policy = "ContentManagePolicy")]
     [HttpGet]
     public async Task<ActionResult<IEnumerable<UserWishListDto>>> GetAll()
     {
@@ -63,6 +70,7 @@ public class UserWishListController : ControllerBase
     /// </summary>
     /// <param name="id">The unique identifier of the UserWishList entry.</param>
     /// <returns>A UserWishListDTO if found.</returns>
+    [Authorize(Policy = "ContentManagePolicy")]
     [HttpGet("{id}")]
     public async Task<ActionResult<UserWishListDto>> GetById([FromRoute] Guid id)
     {
@@ -110,12 +118,13 @@ public class UserWishListController : ControllerBase
     }
 
     /// <summary>
-    /// Adds a new product to the user's wishlist.
+    /// Adds a new item to the user's wish list.
     /// </summary>
-    /// <param name="request">The UserWishListCreateRequest containing user and product IDs.</param>
-    /// <returns>The created UserWishListDTO.</returns>
+    /// <param name="productId">The unique identifier of the product to be added to the wish list.</param>
+    /// <returns>The newly created UserWishListDto containing the details of the added wish list item.</returns>
+    [Authorize]
     [HttpPost]
-    public async Task<ActionResult<UserWishListDto>> Add([FromBody] UserWishListCreateRequest request)
+    public async Task<ActionResult<UserWishListDto>> Add([FromQuery] Guid productId)
     {
         if (!ModelState.IsValid)
         {
@@ -124,7 +133,34 @@ public class UserWishListController : ControllerBase
 
         try
         {
-            var createdWishList = await _userWishListService.AddAsync(request);
+            var keycloakUserId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(keycloakUserId))
+            {
+                return Unauthorized(new { error = "User ID not found in token claims." });
+            }
+
+            var cachedProfile = await _cache.GetStringAsync($"user_profile_{keycloakUserId}");
+            UserProfileDto? userProfile;
+            if (!string.IsNullOrWhiteSpace(cachedProfile))
+            {
+                userProfile = JsonSerializer.Deserialize<UserProfileDto>(cachedProfile);
+            }
+            else
+            {
+                userProfile = await _userService.GetUserByKeycloakIdAsync(keycloakUserId);
+            }
+            
+            if (userProfile == null)
+            {
+                return NotFound(new { error = $"User not found." });
+            }
+            
+            var createdWishList = await _userWishListService.AddAsync(new UserWishListCreateRequest
+            {
+                UserId = userProfile.Id,
+                ProductId = productId
+            });
             if (createdWishList == null)
             {
                 return BadRequest(new { error = "Failed to create wishlist." });
@@ -145,7 +181,7 @@ public class UserWishListController : ControllerBase
                     LogLevel = LogLevel.Error,
                     Details = new Dictionary<string, string>
                     {
-                        { "Message", $"Failed to add user wish lisl for user with id: {request.UserId} and product with id: {request.ProductId}" }
+                        { "Message", $"Failed to add to a wish list product with id: {productId}" }
                     }
                 };
                 
@@ -164,6 +200,7 @@ public class UserWishListController : ControllerBase
     /// </summary>
     /// <param name="request">The UserWishListUpdateRequest containing updated user and product IDs.</param>
     /// <returns>No content if update is successful.</returns>
+    [Authorize(Policy = "ContentManagePolicy")]
     [HttpPut]
     public async Task<ActionResult<UserWishListDto>> Update([FromBody] UserWishListUpdateRequest request)
     {
@@ -214,6 +251,7 @@ public class UserWishListController : ControllerBase
     /// </summary>
     /// <param name="id">The unique identifier of the entry to delete.</param>
     /// <returns>No content if deletion is successful.</returns>
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(Guid id)
     {
@@ -224,6 +262,40 @@ public class UserWishListController : ControllerBase
 
         try
         {
+            var keycloakUserId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(keycloakUserId))
+            {
+                return Unauthorized(new { error = "User ID not found in token claims." });
+            }
+
+            var cachedProfile = await _cache.GetStringAsync($"user_profile_{keycloakUserId}");
+            UserProfileDto? userProfile;
+            if (!string.IsNullOrWhiteSpace(cachedProfile))
+            {
+                userProfile = JsonSerializer.Deserialize<UserProfileDto>(cachedProfile);
+            }
+            else
+            {
+                userProfile = await _userService.GetUserByKeycloakIdAsync(keycloakUserId);
+            }
+            
+            if (userProfile == null)
+            {
+                return NotFound(new { error = $"User not found." });
+            }
+            
+            var wishList = await _userWishListService.GetByIdAsync(id);
+            if (wishList == null)
+            {
+                return NotFound(new { error = $"UserWishList entry with ID {id} not found." });
+            }
+
+            if (wishList.UserId != userProfile.Id)
+            {
+                return Unauthorized(new { error = $"User does not have permission to delete this wish list." });
+            }
+            
             await _userWishListService.DeleteAsync(id);
             return NoContent();
         }
@@ -260,6 +332,7 @@ public class UserWishListController : ControllerBase
     /// </summary>
     /// <param name="userId">The unique identifier of the user.</param>
     /// <returns>A list of UserWishListDTO entries associated with the user.</returns>
+    [Authorize(Policy = "ContentManagePolicy")]
     [HttpGet("user/{userId}")]
     public async Task<ActionResult<IEnumerable<UserWishListDto>>> GetByUserId(Guid userId)
     {
@@ -288,6 +361,66 @@ public class UserWishListController : ControllerBase
                     Details = new Dictionary<string, string>
                     {
                         { "Message", $"Failed to get user wish lisl by user id: {userId}" }
+                    }
+                };
+                
+                await _auditLogService.LogWithLogEntryAsync(logEntry);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine($"Error logging audit log entry: {e.Message}");
+            }
+            return BadRequest(new { error = ex.Message });
+        }
+    }
+    
+    [Authorize]
+    [HttpGet("self")]
+    public async Task<ActionResult<IEnumerable<UserWishListDto>>> GetAllSelfWishLists()
+    {
+        try
+        {
+            var keycloakUserId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(keycloakUserId))
+            {
+                return Unauthorized(new { error = "User ID not found in token claims." });
+            }
+
+            var cachedProfile = await _cache.GetStringAsync($"user_profile_{keycloakUserId}");
+            UserProfileDto? userProfile;
+            if (!string.IsNullOrWhiteSpace(cachedProfile))
+            {
+                userProfile = JsonSerializer.Deserialize<UserProfileDto>(cachedProfile);
+            }
+            else
+            {
+                userProfile = await _userService.GetUserByKeycloakIdAsync(keycloakUserId);
+            }
+            
+            if (userProfile == null)
+            {
+                return NotFound(new { error = $"User not found." });
+            }
+            
+            var wishLists = await _userWishListService.GetByUserIdWithExtendedGameAsync(userProfile.Id);
+            return Ok(wishLists);
+        }
+        catch (Exception ex)
+        {
+            try
+            {
+                var logEntry = new LogEntry<string>
+                {
+                    State = "Error",
+                    UserId = null,
+                    Exception = ex,
+                    ActionType = AuditActionType.View,
+                    EntityType = AuditEntityType.Other,
+                    LogLevel = LogLevel.Error,
+                    Details = new Dictionary<string, string>
+                    {
+                        { "Message", $"Failed to get user self wish lisl for user" }
                     }
                 };
